First patch of a serie of two correcting a fftw related bug.
Bug description: 
SFXC use the fftw Guru interface (for high performance) this means that 
some care have to be taken on the data passed 
to fftw (same size and SAME ALIGNMENT). There is no easy way to 
force a std::vector &(data[0]) to be correctly aligned so it 
may crash the application on some system during fftw. 
 
To fix this I use now fft_malloc/free to allocate an array 
correctly aligned. 



git-svn-id: svn+ssh://jop91.astron.nl/auto/home/jive_cvs/SOFTC/svn/trunk/sfxc@578 5736e3ea-d426-0410-8aaf-93e336ac3372