Add the recording of monitoring info into the correlation_node. If sfxc is compiled with -DRUNTIME_STATISTIC be sure to have enough disk-space when you run SFXC as 5 samples are produced per second. And each node have around 5 monitors.
All the related code is deactivated with -DRUNTIME_STATISTIC. 

This version was tested for regression by correlating one second
of n06c2. Result are the same as with r719. 

This version was tested for regression by correlating one minute
of n06c2 on DAS3. The fringes look like fringes. Speed 
is around 30% of realtime.




git-svn-id: svn+ssh://jop91.astron.nl/auto/home/jive_cvs/SOFTC/svn/trunk/sfxc@729 5736e3ea-d426-0410-8aaf-93e336ac3372